cast it into unsigned long

+++ libvpx_patches/source/libvpx/vp8/encoder/encodeframe.c
@@ -951,7 +951,7 @@ void vp8_encode_frame(VP8_COMP *cpi) {
     end = clock();
     //printf("Time spent generating tokens %d: %lf\n", cm->current_video_frame, ((double) end - start) / CLOCKS_PER_SEC);
     //Stegozoa: embedding capacity
-    fprintf(stdout, "Ssrc: %lu. Current Frame: %d. Embbed data: %d\n", cpi->ssrc, cpi->common.current_video_frame, embbedData);
+    fprintf(stdout, "Ssrc: %lu. Current Frame: %d. Embbed data: %d\n", (unsigned long) cpi->ssrc, cpi->common.current_video_frame, embbedData);
     fflush(stdout);
